The House Sale Price Prediction with the oneAPI
We are trying to predict the house sale price based on different features like zone, area, year built, etc. We fetch data, clean and build regression to predict the house sale values. We used intel oneAPI features to build this. There are 37 different features that we considered to generate the sale prediction. When used with intel optimized packages 11% speed increase is seen also mean absolute error reduced by 1%. Have a look at the code and try it out! it's fun!